user-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
authlib==1.3.2
python-dotenv==1.0.1
httpx==0.27.2  # Added to resolve the dependency
itsdangerous==2.2.0  # Added for 
PyJWT>=2.8.0
kafka-python==2.0.2
